requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alive-progress==3.3.0
aniso8601==10.0.1
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
autograd==1.8.0
bayesian-optimization==1.4.0
blinker==1.9.0
booz_xform==0.0.9
certifi
cftime==1.6.5
charset-normalizer==3.4.4
choreographer==1.2.0
click==8.3.0
cma==4.4.0
comm
constellaration==0.2.4
contourpy==1.3.3
coptpy==7.2.11
cycler==0.12.1
datasets==3.5.1
debugpy
decorator==5.2.1
Deprecated==1.2.18
desc-opt==0.15.0
diffrax==0.7.0
dill==0.3.8
directsearch==1.0
distro==1.9.0
einops==0.8.1
equinox==0.12.2
executing==2.2.1
f90nml==1.5.0
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
Flask-RESTful==0.3.10
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
graphemeu==0.7.2
gurobipy==12.0.3
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hvwfg==1.0.2
idna==3.11
imageio==2.37.0
iniconfig==2.3.0
interpax==0.3.8
ipykernel==7.0.1
iteration_utilities==0.13.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
jaxtyping==0.3.3
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
kaleido==1.1.0
kiwisolver==1.4.9
lazy_loader==0.4
lineax==0.0.8
logistro==2.0.0
makefun==1.16.0
MarkupSafe==3.0.3
matplotlib==3.9.2
mkl-service==2.5.2
mkl_fft==2.0.0
mkl_random
ml_dtypes==0.5.3
monty==2025.3.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.9.0
nbformat==5.10.4
nest_asyncio
netCDF4==1.7.2
networkx==3.5
nevergrad==1.0.12
numexpr
numpy
nvgpu==0.10.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.3.0
opt_einsum==3.4.0
optimistix==0.0.11
orjson==3.10.16
orthax==0.2.5
packaging
pandas==2.2.3
pathspec==0.12.1
pillow==12.0.0
plotly==6.3.0
pluggy==1.6.0
propcache==0.4.1
psutil==7.0.0
pyarrow==21.0.0
pybind11==2.13.6
pydantic==2.12.3
pydantic_core==2.41.4
pyevtk==1.6.0
pylatexenc==2.10
pymoo==0.6.1.5.dev0
pynvml==13.0.1
pyparsing==3.2.5
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz
PyYAML==6.0.3
pyzmq==27.1.0
qicna==0.3.4
quadax==0.2.9
randomgen==2.3.0
rdkit==2025.9.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-activeml==0.5.2
scikit-image==0.25.2
scikit-learn==1.5.2
scikit_build_core==0.10.0
scipy
simplejson==3.20.2
simsopt==1.8.1
six
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.8.0
tqdm==4.67.1
traitlets
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata
urllib3==2.5.0
vmecpp==0.4.7
wadler_lindig==0.1.7
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
